// src/api/model/dashboard.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Widget } from './widget';

/**
 * Dashboard for a user
 */
export interface Dashboard { 
    createdAt?: string;
    createdBy?: number;
    id: number;
    /**
     * Dashboard name
     */
    name: string;
    /**
     * Group foreign key
     */
    groupId?: number;
    /**
     * User foreign key
     */
    userId: number;
    updatedAt?: string;
    /**
     * Widgets associated to dashboard
     */
    widgets?: Array<Widget>;
}